+++ crius-service/src/main/java/com/magic/crius/service/UserInfoService.java
@@ -13,5 +13,5 @@ public interface UserInfoService {
 
     UserInfo get(Long userId);
 
-    boolean updateLevel(Long userId, Integer level);
+    boolean updateLevel(Long userId, Long level);
 }
